+++ src/main/java/org/libreoffice/ci/gerrit/buildbot/model/TBBlockingQueue.java
@@ -9,9 +9,12 @@
 
 package org.libreoffice.ci.gerrit.buildbot.model;
 
+import java.io.PrintWriter;
 import java.io.Serializable;
 import java.util.LinkedList;
 import java.util.Set;
+
+import org.libreoffice.ci.gerrit.buildbot.utils.QueueUtils;
 
 import com.google.common.collect.Lists;
 
@@ -72,6 +75,18 @@ public class TBBlockingQueue implements Serializable {
         synchronized (queue) {
             return queue.isEmpty();
         }
+    }
+
+    public int dumpTasks(PrintWriter stdout) {
+        int pendingTasks = 0;
+        synchronized (queue) {
+            if (queue.isEmpty()) {
+                stdout.print("empty\n");
+                return pendingTasks;
+            }
+            pendingTasks = QueueUtils.dumpTasks(stdout, pendingTasks, queue);
+        }
+        return pendingTasks;
     }
 
 }